Uniform mutator config
